Server-Side Login/Register Update
Edited DataRepository to create a user if passed a boolean of true as the create parameter. Also sets up a bank account that gets attached to the user and adds 1000 credits if the account was newly created. The method returns the existing account otherwise. AccountController calls this method for both of its Wrapper cases of register and login.