Fixed Movement - Needs Further Work
I fixed movement to have the left and right buttons turn the tank either direction and the up and down buttons work to move forward or backward based on the direction of the tank. I created a controller class per Prof. Plumlee to handle these, and I will need to update both the new Controller class and the ClientActivity to move all cases of the RestClient's usage out of the ClientActivity and into the new Controller class (TankController). Can expect to have all of this done by the end of today (Monday). I need to implement testing as well as Javadoc for everything I've moved/written so far. Can expect to have that done by mid-to-end of tomorrow.