Fully Implemented Controller/Javadoc Implementation
I have fully implemented the Controller for the ClientActivity (TankController), and have written all the Javadoc for it as well. Everything works in terms of the app running and functioning as expected. I also edited the ClientActivity for these purposes and fully implemented the ShakeSensor to fire the bullet. I made very small changes to the account handling on the server side in terms of Javadoc as well. I will work on the client-side implementation now with Anthony's changes to the server.